CHG: Move summary statistics into own module, remove methods
- so now we always have `spy.mean(data, ...)` and so on

Changes to be committed:
	modified:   syncopy/datatype/__init__.py
	modified:   syncopy/datatype/continuous_data.py
	modified:   syncopy/statistics/__init__.py
	renamed:    syncopy/datatype/methods/statistics.py -> syncopy/statistics/summary_stats.py
	modified:   syncopy/tests/test_continuousdata.py
	new file:   syncopy/tests/test_statistics.py
